chore(landing): separate buttons into individual components to render on client
