+++ tavla/app/components/CreateUserButtonLanding/index.tsx
@@ -1,0 +1,22 @@
+'use client'
+import { Button } from '@entur/button'
+import Link from 'next/link'
+import { usePostHog } from 'posthog-js/react'
+
+function CreateUserButtonLanding() {
+    const posthog = usePostHog()
+    return (
+        <Button
+            variant="success"
+            size="medium"
+            as={Link}
+            href="?login=create"
+            onClick={() => {
+                posthog.capture('CREATE_USER_BTN_FROM_LANDING')
+            }}
+        >
+            Opprett bruker
+        </Button>
+    )
+}
+export { CreateUserButtonLanding }
